requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HuggingFace integration
huggingface-hub==0.19.4
transformers==4.35.2

# Image processing
Pillow==10.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Async support
asyncio==3.4.3

# Logging and monitoring
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0
